web.common: renamed Heading component to Headline
