+++ products/ASC.People/Client/src/components/pages/Profile/Section/Header/index.js
@@ -12,7 +12,7 @@ import {
   ModalDialog,
   AvatarEditor,
 } from "asc-web-components";
-import { Heading } from 'asc-web-common';
+import { Headline } from 'asc-web-common';
 import { withRouter } from "react-router";
 import {
   getUserStatus,
@@ -44,7 +44,7 @@ const wrapperStyle = {
   alignItems: "center"
 };
 
-const HeaderContainer = styled(Heading)`
+const HeaderContainer = styled(Headline)`
   margin-left: 16px;
   margin-right: 16px;
   max-width: calc(100vw - 430px);
