+++ web/ASC.Web.Client/src/components/pages/Settings/Layout/Article/Header/index.js
@@ -1,10 +1,10 @@
 import React from 'react';
-import { Heading } from 'asc-web-common';
+import { Headline } from 'asc-web-common';
 import { useTranslation } from 'react-i18next';
 
 const ArticleHeaderContent = () => {
   const { t } = useTranslation();
-  return <Heading type="menu">{t('Settings')}</Heading>;
+  return <Headline type="menu">{t('Settings')}</Headline>;
 }
 
 export default ArticleHeaderContent;
