+++ web/ASC.Web.Client/src/components/pages/Settings/Layout/Section/Header/index.js
@@ -1,12 +1,12 @@
 import React from "react";
 import { withRouter } from "react-router";
 import { utils } from 'asc-web-components';
-import { Heading } from 'asc-web-common';
+import { Headline } from 'asc-web-common';
 import styled from 'styled-components';
 import { withTranslation } from 'react-i18next';
 import { getKeyByLink, settingsTree, getTKeyByKey } from '../../../utils';
 
-const HeaderContainer = styled(Heading)`
+const HeaderContainer = styled(Headline)`
   margin-right: 16px;
   max-width: calc(100vw - 430px);
   @media ${utils.device.tablet} {
