+++ web/ASC.Web.Common/src/components/headline/index.js
@@ -1,7 +1,7 @@
 import React from "react";
 import PropTypes from 'prop-types';
 import styled, { css } from 'styled-components';
-import { Heading as PureHeading } from 'asc-web-components';
+import { Heading } from 'asc-web-components';
 
 const fontSize = css`
       ${props =>
@@ -10,22 +10,22 @@ const fontSize = css`
    }
    `;
 
-const StyledHeading = styled(PureHeading)`
+const StyledHeading = styled(Heading)`
    margin: 0;
    line-height: 56px;
    font-size: ${fontSize}px;
    font-weight: 700;
 `;
 
-const Heading = (props) => {
-   //console.log("Heading render");
+const Headline = (props) => {
+   //console.log("Headline render");
    return (
       <StyledHeading
          {...props} />
    );
 };
 
-Heading.propTypes = {
+Headline.propTypes = {
    level: PropTypes.oneOf([1, 2, 3, 4, 5, 6]),
    children: PropTypes.any,
    color: PropTypes.string,
@@ -35,7 +35,7 @@ Heading.propTypes = {
    type: PropTypes.oneOf(['menu', 'content']),
 };
 
-Heading.defaultProps = {
+Headline.defaultProps = {
    color: '#333333',
    title: '',
    truncate: false,
@@ -43,4 +43,4 @@ Heading.defaultProps = {
    level: 1
 };
 
-export default Heading;
+export default Headline;
