+++ web/ASC.Web.Common/src/components/index.js
@@ -3,4 +3,4 @@ export { default as PublicRoute } from "./routing/publicRoute";
 export { default as Login } from "./login";
 export { default as StudioLayout } from "./layout";
 export { default as ModuleTile } from "./module-tile";
-export { default as Heading } from "./heading";
+export { default as Headline } from "./headline";
